+++ src/stylesheet/dark/navigation_bar_data.h
@@ -12,24 +12,24 @@
 extern "C" {
 #endif
 
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_back);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_home);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_recents_screen);
+LV_IMG_DECLARE(esp_ui_phone_320_240_navigation_bar_image_back);
+LV_IMG_DECLARE(esp_ui_phone_320_240_navigation_bar_image_home);
+LV_IMG_DECLARE(esp_ui_phone_320_240_navigation_bar_image_recents_screen);
 
-#define ESP_UI_PHONE_[T_WIDTH]_[T_HEIGHT]_DARK_NAVIGATION_BAR_DATA()                                              \
+#define ESP_UI_PHONE_320_240_DARK_NAVIGATION_BAR_DATA()                                              \
     {                                                                                                \
         .main = {                                                                                    \
-            .size = ESP_UI_STYLE_SIZE_RECT_PERCENT(100, 15),                                         \
-            .size_min = ESP_UI_STYLE_SIZE_RECT_W_PERCENT(100, 36),                                   \
-            .size_max = ESP_UI_STYLE_SIZE_RECT_W_PERCENT(100, 60),                                   \
+            .size = ESP_UI_STYLE_SIZE_RECT_W_PERCENT(100, 40),                                       \
+            .size_min = {},                                                                          \
+            .size_max = {},                                                                          \
             .background_color = ESP_UI_STYLE_COLOR(0x38393A),                                        \
         },                                                                                           \
         .button = {                                                                                  \
-            .icon_size = ESP_UI_STYLE_SIZE_SQUARE_PERCENT(50),                                       \
+            .icon_size = ESP_UI_STYLE_SIZE_SQUARE_PERCENT(60),                                       \
             .icon_images = {                                                                         \
-                ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_back),           \
-                ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_home),           \
-                ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_recents_screen), \
+                ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_navigation_bar_image_back),           \
+                ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_navigation_bar_image_home),           \
+                ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_navigation_bar_image_recents_screen), \
             },                                                                                       \
             .navigate_types = {                                                                      \
                 ESP_UI_CORE_NAVIGATE_TYPE_BACK,                                                      \
@@ -48,8 +48,8 @@ LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_navigation_bar_image_recents_sc
             .hide_animation_path_type = ESP_UI_LV_ANIM_PATH_TYPE_EASE_IN,                            \
         },                                                                                           \
         .flags = {                                                                                   \
-            .enable_main_size_min = 1,                                                               \
-            .enable_main_size_max = 1,                                                               \
+            .enable_main_size_min = 0,                                                               \
+            .enable_main_size_max = 0,                                                               \
         },                                                                                           \
     }