+++ src/stylesheet/dark/status_bar_data.h
@@ -14,56 +14,56 @@ extern "C" {
 
 /* Icon Images */
 // Battery
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_charge);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_1);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_2);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_3);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_4);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_battery_image_charge);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_battery_image_level_1);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_battery_image_level_2);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_battery_image_level_3);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_battery_image_level_4);
 // WiFi
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_close);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_1);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_2);
-LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_3);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_wifi_image_close);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_wifi_image_level_1);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_wifi_image_level_2);
+LV_IMG_DECLARE(esp_ui_phone_320_240_status_bar_wifi_image_level_3);
 
 /* Area */
-#define ESP_UI_PHONE_[T_WIDTH]_[T_HEIGHT]_DARK_STATUS_BAR_AREA_DATA(w_percent, align) \
+#define ESP_UI_PHONE_320_240_DARK_STATUS_BAR_AREA_DATA(w_percent, align) \
     {                                                                    \
         .size = ESP_UI_STYLE_SIZE_RECT_PERCENT(w_percent, 100),          \
         .layout_column_align = align,                                    \
-        .layout_column_start_offset = 10,                                \
+        .layout_column_start_offset = 12,                                \
         .layout_column_pad = 3,                                          \
     }
 
 /* Status Bar */
-#define ESP_UI_PHONE_[T_WIDTH]_[T_HEIGHT]_DARK_STATUS_BAR_DATA()                                                       \
+#define ESP_UI_PHONE_320_240_DARK_STATUS_BAR_DATA()                                                       \
     {                                                                                                     \
         .main = {                                                                                         \
-            .size = ESP_UI_STYLE_SIZE_RECT_PERCENT(100, 10),                                              \
-            .size_min = ESP_UI_STYLE_SIZE_RECT_W_PERCENT(100, 24),                                        \
-            .size_max = ESP_UI_STYLE_SIZE_RECT_W_PERCENT(100, 50),                                        \
+            .size = ESP_UI_STYLE_SIZE_RECT_W_PERCENT(100, 36),                                            \
+            .size_min = {},                                                                               \
+            .size_max = {},                                                                               \
             .background_color = ESP_UI_STYLE_COLOR(0x38393A),                                             \
-            .text_font = ESP_UI_STYLE_FONT_HEIGHT_PERCENT(60),                                            \
+            .text_font = ESP_UI_STYLE_FONT_HEIGHT_PERCENT(56),                                            \
             .text_color = ESP_UI_STYLE_COLOR(0xFFFFFF),                                                   \
         },                                                                                                \
         .area = {                                                                                         \
             .num = 2,                                                                                     \
             .data = {                                                                                     \
-                ESP_UI_PHONE_[T_WIDTH]_[T_HEIGHT]_DARK_STATUS_BAR_AREA_DATA(50, ESP_UI_STATUS_BAR_AREA_ALIGN_START),   \
-                ESP_UI_PHONE_[T_WIDTH]_[T_HEIGHT]_DARK_STATUS_BAR_AREA_DATA(50, ESP_UI_STATUS_BAR_AREA_ALIGN_END),     \
+                ESP_UI_PHONE_320_240_DARK_STATUS_BAR_AREA_DATA(50, ESP_UI_STATUS_BAR_AREA_ALIGN_START),   \
+                ESP_UI_PHONE_320_240_DARK_STATUS_BAR_AREA_DATA(50, ESP_UI_STATUS_BAR_AREA_ALIGN_END),     \
             },                                                                                            \
         },                                                                                                \
-        .icon_common_size = ESP_UI_STYLE_SIZE_SQUARE_PERCENT(60),                                         \
+        .icon_common_size = ESP_UI_STYLE_SIZE_SQUARE_PERCENT(56),                                         \
         .battery = {                                                                                      \
             .area_index = 1,                                                                              \
             .icon_data = {                                                                                \
                 .icon = {                                                                                 \
                     .image_num = 5,                                                                       \
                     .images = {                                                                           \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_1), \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_2), \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_3), \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_level_4), \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_battery_charge),  \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_battery_image_level_1), \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_battery_image_level_2), \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_battery_image_level_3), \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_battery_image_level_4), \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_battery_image_charge),  \
                     },                                                                                    \
                 },                                                                                        \
             },                                                                                            \
@@ -74,10 +74,10 @@ LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_3);
                 .icon = {                                                                                 \
                     .image_num = 4,                                                                       \
                     .images = {                                                                           \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_close),      \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_1),    \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_2),    \
-                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_3),    \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_wifi_image_close),      \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_wifi_image_level_1),    \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_wifi_image_level_2),    \
+                        ESP_UI_STYLE_IMAGE_RECOLOR_WHITE(&esp_ui_phone_320_240_status_bar_wifi_image_level_3),    \
                     },                                                                                    \
                 },                                                                                        \
             },                                                                                            \
@@ -86,8 +86,8 @@ LV_IMG_DECLARE(esp_ui_phone_[T_WIDTH]_[T_HEIGHT]_status_bar_image_wifi_level_3);
             .area_index = 0,                                                                              \
         },                                                                                                \
         .flags = {                                                                                        \
-            .enable_main_size_min = 1,                                                                    \
-            .enable_main_size_max = 1,                                                                    \
+            .enable_main_size_min = 0,                                                                    \
+            .enable_main_size_max = 0,                                                                    \
             .enable_battery_icon = 1,                                                                     \
             .enable_battery_icon_common_size = 1,                                                         \
             .enable_battery_label = 0,                                                                    \